refactors auth components
