+++ src/components/Auth/ResetPassword/index.js
@@ -1,0 +1,4 @@
+import ResetPasswordForm from './ResetPasswordForm';
+import ResetPasswordLink from './ResetPasswordLink';
+
+export { ResetPasswordForm, ResetPasswordLink };
