+++ src/components/Auth/index.js
@@ -1,15 +1,9 @@
-import Signup from './Signup';
-import Login from './Login';
 import Logout from './Logout';
 import withAuthentication from './withAuthentication';
 import withAuthorisation from './withAuthorisation';
-import ResetPasswordForm from './ResetPasswordForm';
 
 export {
-  Signup,
-  Login,
   Logout,
   withAuthentication,
-  withAuthorisation,
-  ResetPasswordForm
+  withAuthorisation
 };
